requirements.txt
# Core
pyyaml>=6.0.1,<7.0.0
python-dotenv>=1.0.1,<2.0.0
loguru>=0.7.2,<0.8.0
pydantic>=2.5.0,<3.0.0
dataclasses-json>=0.6.3,<1.0.0
python-dateutil>=2.8.2,<3.0.0

# HTTP client (for OpenClaw Gateway API)
httpx>=0.25.0,<1.0.0
playwright>=1.52.0,<2.0.0
websockets>=12.0,<16.0
python-socks[asyncio]>=2.4.0,<3.0.0

# Image Processing
Pillow>=10.2.0,<11.0.0
opencv-python-headless>=4.8.0,<5.0.0

# Browser stealth
playwright-stealth>=2.0.0,<3.0.0

# OSS Storage
oss2>=2.18.0,<3.0.0

# Async / Database
aiosqlite>=0.19.0,<1.0.0
aiofiles>=23.2.1,<24.0.0
pandas>=2.1.0,<3.0.0

# Cookie auto-grab from browser DB
rookiepy>=0.5.0,<1.0.0

# Encryption
cryptography>=42.0.0,<44.0.0

# AI / LLM
openai>=1.12.0,<2.0.0

# Testing baseline
pytest>=7.4.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
pytest-cov>=4.1.0,<6.0.0
